Fix bug in docstring parsing, docstrings inherit from bases (#164)
* Add test to reproduce issue 162


* Use new style types in docstring.py


* Fix bug in get_attribute_docstring


* Docstring parts are inherited from subclasses


* Fix weird `inspect.getsource` bug in py37


* Reduce unfound docstring message to debug level


* use type.__name__ in log message


* Remove numpy warning for use of np.int


* Add issue link in test docstring

+++ simple_parsing/helpers/hparams/hyperparameters_test.py
@@ -265,6 +265,6 @@ def test_field_types():
     assert C.get_priors()["f"].discrete is True
 
     if numpy_installed:
-        assert all(c.f.dtype == np.int for c in cs)
+        assert all(c.f.dtype == int for c in cs)
     else:
         assert all(all(isinstance(v, int) for v in c.f) for c in cs)
